fix: uint32 overflow in FlatGraph seq_base_offset (silent corruption past 4 GB)
When total bases (forward + reverse, summed across all nodes) exceeds
4 GB, the per-node seq_base_offset entries silently overflowed,
pointing later nodes at WRONG locations in the packed sequence arena.
Walking those nodes during indexing read garbage bases -> wrong kmers
-> wrong hashes inserted under those node_ids. Lookups for the real
cluster minimizers returned nothing.

Concretely on hg38 (3 GB * 2 = ~6.2 GB packed): chr14 onward had
seq_base_offset wrapped from ~4.4 GB to ~83 MB. Indexing chr14_rev
read bytes belonging to chr1's packed region. The chr14 cluster
minimizers (e.g., hash 175201191 at chr14_rev:19267650) never made
it into the index. dev-108 G5 NA12878 500-read mapping topped out at
~258 instead of minimap2's 378.

Fix: change seq_base_offset_ to uint64 in FlatGraph + propagate through
simpleExpandFlat (base_cursor was already size_t but cast to uint32 on
store) and base2bit/isN (abs_pos / byte_idx widened to uint64).

Side benefit: byte_idx in base2bit/isN can now address arenas > 4 GB,
so packed_bytes is no longer the implicit ceiling either.

Also dropped the temporary PIRU_DUMP_LOOKUPS debug instrumentation
and a stray blank line in BaseSeedLookup::lookup.

Verification on hg38 NA12878 500 reads (rebuilt index after fix):
  pre-fix default:           209 mapped
  pre-fix combo:             258 mapped
  post-fix default:          323 mapped (+114 vs pre-fix)
  post-fix combo:            360 mapped (+102 vs pre-fix combo)
  minimap2 ground truth:     378

96% of minimap2's recall now, vs ~68% before. Specificity holds 1.000
on ecoli off-target. Covid + yeast unchanged (small graphs, never
triggered the overflow). All ctest pass except the 2 pre-existing
blow5-fixture failures.

Existing .pirx files built from references > 4 GB total bases
(forward + reverse) MUST BE REBUILT after this fix.

+++ include/core/index/flat_graph.hpp
@@ -46,7 +46,7 @@ public:
   static FlatGraph fromPackedArrays(
       std::uint32_t node_count, std::uint32_t path_count, std::size_t total_bases,
       std::vector<std::uint8_t> seq_packed, std::vector<std::uint8_t> seq_n_mask,
-      std::vector<std::uint32_t> seq_base_offset, std::vector<std::uint32_t> seq_len,
+      std::vector<std::uint64_t> seq_base_offset, std::vector<std::uint32_t> seq_len,
       std::vector<char> name_data, std::vector<std::uint32_t> name_offset,
       std::vector<std::uint16_t> name_len, std::vector<std::uint8_t> is_reverse,
       std::vector<std::uint32_t> edge_target, std::vector<std::uint32_t> out_edge_offset,
@@ -85,17 +85,19 @@ public:
 
   std::size_t seqLen(std::uint32_t node_id) const { return seq_len_[node_id]; }
 
-  // Get 2-bit encoded base at position within node (0-3, check isN for ambiguity)
+  // Get 2-bit encoded base at position within node (0-3, check isN for ambiguity).
+  // abs_pos / byte_idx must be 64-bit: total_bases on hg38 (fwd+rev) is ~6.2 GB,
+  // overflowing uint32. (Bug fixed dev-108 -- silent corruption past 4 GB cumulative.)
   std::uint8_t base2bit(std::uint32_t node_id, std::uint32_t pos) const {
-    std::uint32_t abs_pos = seq_base_offset_[node_id] + pos;
-    std::uint32_t byte_idx = abs_pos >> 2;         // abs_pos / 4
+    std::uint64_t abs_pos = seq_base_offset_[node_id] + pos;
+    std::uint64_t byte_idx = abs_pos >> 2;         // abs_pos / 4
     std::uint32_t bit_shift = (abs_pos & 3) << 1;  // (abs_pos % 4) * 2
     return (seq_packed_[byte_idx] >> bit_shift) & 3;
   }
 
   // Check if position is an N base
   bool isN(std::uint32_t node_id, std::uint32_t pos) const {
-    std::uint32_t abs_pos = seq_base_offset_[node_id] + pos;
+    std::uint64_t abs_pos = seq_base_offset_[node_id] + pos;
     return (seq_n_mask_[abs_pos >> 3] >> (abs_pos & 7)) & 1;
   }
 
@@ -165,7 +167,7 @@ public:
 
   const std::vector<std::uint8_t>& seqPacked() const { return seq_packed_; }
   const std::vector<std::uint8_t>& seqNMask() const { return seq_n_mask_; }
-  const std::vector<std::uint32_t>& seqBaseOffsets() const { return seq_base_offset_; }
+  const std::vector<std::uint64_t>& seqBaseOffsets() const { return seq_base_offset_; }
   const std::vector<std::uint32_t>& seqLens() const { return seq_len_; }
   void setSeqLens(std::vector<std::uint32_t> lens) { seq_len_ = std::move(lens); }
   std::size_t totalBases() const { return total_bases_; }
@@ -189,7 +191,7 @@ private:
   // 2-bit packed sequence arena (4 bases per byte)
   std::vector<std::uint8_t> seq_packed_;        // ceil(total_bases / 4) bytes
   std::vector<std::uint8_t> seq_n_mask_;        // 1 bit per base (ceil(total_bases / 8) bytes)
-  std::vector<std::uint32_t> seq_base_offset_;  // [node_count] offset in bases
+  std::vector<std::uint64_t> seq_base_offset_;  // [node_count] offset in bases
   std::vector<std::uint32_t> seq_len_;          // [node_count] length in bases
 
   // Name arena (shared by nodes and paths)
